Add passwords router and export post and get function from server.js
